requirements.txt
# Core packages for lecture video transcription
openai-whisper>=20231117
opencv-python>=4.8.0
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
plotly>=5.15.0

# Additional useful packages for video/audio processing
ffmpeg-python>=0.2.0
librosa>=0.10.0
soundfile>=0.12.0
torch>=2.0.0
torchaudio>=2.0.0

# Development and utility packages
jupyter>=1.0.0
ipykernel>=6.25.0 